Move logging related classes to a separate module
