+++ modules/ballerina-logging/src/main/java/org/ballerinalang/logging/BLogManager.java
@@ -6,7 +6,7 @@
  *  in compliance with the License.
  *  You may obtain a copy of the License at
  *
- *  http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  *  Unless required by applicable law or agreed to in writing,
  *  software distributed under the License is distributed on an
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.launcher;
+package org.ballerinalang.logging;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
